+++ Sources/VComponents/Components/Containers/Disclosure Group/VDisclosureGroupLabel.swift
@@ -1,5 +1,5 @@
 //
-//  VAccordionHeaderLabel.swift
+//  VDisclosureGroupLabel.swift
 //  VComponents
 //
 //  Created by Vakhtang Kontridze on 4/6/22.
@@ -7,8 +7,8 @@
 
 import SwiftUI
 
-// MARK: - V Accordion Header Label
-enum VAccordionHeaderLabel<CustomHeaderLabel> where CustomHeaderLabel: View {
+// MARK: - V Disclosure Group Header Label
+enum VDisclosureGroupLabel<CustomHeaderLabel> where CustomHeaderLabel: View {
     case title(title: String)
     case custom(label: () -> CustomHeaderLabel)
 }